product return endpoint

+++ src/main/java/ma/nemo/assignment/service/ProductService.java
@@ -7,9 +7,9 @@ import ma.nemo.assignment.domain.Product;
 import ma.nemo.assignment.domain.ProductModel;
 
 public interface ProductService {
-    Product createProductWith(ProductModel productModel, Integer quantity);
+    Product createProductWith(ProductModel productModel, long quantity);
 
-    Product createProductWith(ProductModel productModel, Integer quantity, Date expirationDate);
+    Product createProductWith(ProductModel productModel, long quantity, Date expirationDate);
 
     List<Product> getNearingExpirationDateProducts();
